requirements.txt
requests==2.31.0
flask==3.0.0
Pillow==8.3.2
facebook-sdk==3.0
